seed/app/src/main/java/com/WATER_ME_WITH_COMPANY_GOODNESS/android/WATER_ME_WITH_APP_NAME/util/ViewUtils.java
package com.{company_name}.android.{app_package_name_prefix}.util;

import android.view.View;

/**
 *
 */
public class ViewUtils {

    public static void hide(View... views) {
        if (views != null) {
            for (View view : views) {
                if (view != null) {
                    view.setVisibility(View.GONE);
                }
            }
        }
    }

    public static void show(View... views) {
        if (views != null) {
            for (View view : views) {
                if (view != null) {
                    view.setVisibility(View.VISIBLE);
                }
            }
        }
    }
}
